requirements.txt
apns2==0.7.1
asgiref==3.3.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
dj-static==0.0.6
Django==3.0.8
django-appconf==1.0.4
django-filter==2.4.0
django-rest-framework==0.1.0
django-rest-knox==4.1.0
django-stdimage==5.1.1
djangorestframework==3.12.1
drf-writable-nested==0.6.1
h2==4.0.0
hpack==4.0.0
http-ece==1.1.0
hyper==0.7.0
hyperframe==6.0.0
idna==2.10
ldif3==3.2.2
Pillow==8.0.1
progressbar2==3.53.1
psycopg2==2.8.6
py-vapid==1.7.1
pycparser==2.20
PyJWT==1.7.1
python-card-me==0.9.3
python-dateutil==2.8.1
python-utils==2.4.0
pytz==2020.1
pywebpush==1.11.0
requests==2.24.0
six==1.15.0
sorl-thumbnail==12.6.3
sqlparse==0.4.1
static3==0.7.0
surlex==0.2.0
unicodecsv==0.14.1
urllib3==1.25.11
